% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density.R
\name{calculate_density}
\alias{calculate_density}
\title{Calculate Density of Numeric Vector}
\usage{
calculate_density(x, bw = "nrd0", adjust = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{bw}{Bandwidth for density calculation; default is "nrd0".}

\item{adjust}{Adjustment factor for the bandwidth; default is 1.}
}
\value{
A list containing density values, input summary, and x values.
}
\description{
Computes the density of a given numeric vector using kernel density estimation.
The result includes the density values, summary statistics, and x-values for the plot.
}
\examples{
x <- rnorm(100)
result <- calculate_density(x)
}
