% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density Data}
\usage{
plot_density(density_data, main = "Density Plot")
}
\arguments{
\item{density_data}{List containing the results of `calculate_density`.}

\item{main}{Title for the plot.}
}
\value{
Plot of the density data.
}
\description{
This function plots the density of data calculated by `calculate_density`.
}
\examples{
data <- rnorm(100)
density_result <- calculate_density(data)
plot_density(density_result)
}
